Display trip headsigns as the follow-up options
Instead of presenting the user with a list of other stops near the given
location, we now present a list of trip headsigns. The nearby stops are
often ambiguously named, so it's not clear what to choose. Especially in
those ambiguous cases, one stop might have the northbound routes, and
the other might have the southbound routes. So we present those
distinguishable and relevant options.